pyproject.toml
[project]
name = "face-regcognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "face-recognition>=1.3.0",
    "opencv-python>=4.10.0.84",
    "pillow>=11.0.0",
    "ultralytics>=8.3.47",
]
